+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Improve accuracy of resource attributes `telemetry.distro.name` and `telemetry.distro.version`.
+
 ## 0.6.0-beta.1
 
 Released 2023-11-09